reorganization of the generic module
